chore: Cleanup code / make all internal by default
