+++ src/Uno.Extensions.Core/Option.cs
@@ -21,12 +21,6 @@ public static class Option
 	public static Option<T> Undefined<T>() => Option<T>.Undefined();
 }
 
-//public readonly struct ListOption<T>
-//{
-//	public static implicit operator Option<IImmutableList<T>>(ListOption<T>? value)
-//		=> new(OptionType.Some, value.);
-//}
-
 public readonly struct Option<T> : IOption, IEquatable<Option<T>>
 {
 	public static Option<T> Some(T? value)
