+++ src/Uno.Extensions.Reactive.UI/Presentation/Bindings/Collections/Data/DataLayerTracker.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using Uno.Extensions.Collections;
 using Uno.Extensions.Collections.Tracking;
 
-namespace Umbrella.Presentation.Feeds.Collections._BindableCollection.Data
+namespace Uno.Extensions.Reactive.Bindings.Collections._BindableCollection.Data
 {
 	internal sealed class DataLayerTracker : ILayerTracker
 	{
