+++ src/Uno.Extensions.Reactive.UI/Presentation/Bindings/Collections/Data/IUpdateContext.cs
@@ -3,7 +3,7 @@ using System.Collections.Specialized;
 using System.Linq;
 using Uno.Extensions.Collections;
 
-namespace Umbrella.Presentation.Feeds.Collections._BindableCollection.Data
+namespace Uno.Extensions.Reactive.Bindings.Collections._BindableCollection.Data
 {
 	/// <summary>
 	/// A counter that can be used to configure the tracking behavior
