+++ src/Uno.Extensions.Reactive.UI/Presentation/Bindings/Collections/VectorChangedEventArgs.cs
@@ -2,23 +2,22 @@
 using System.Linq;
 using Windows.Foundation.Collections;
 
-namespace Uno.Extensions.Collections
+namespace Uno.Extensions.Reactive.Bindings.Collections;
+
+/// <summary>
+/// Default implementation of <see cref="IVectorChangedEventArgs"/>.
+/// </summary>
+internal class VectorChangedEventArgs : IVectorChangedEventArgs
 {
-	/// <summary>
-	/// Default implementation of <see cref="IVectorChangedEventArgs"/>.
-	/// </summary>
-	public class VectorChangedEventArgs : IVectorChangedEventArgs
+	public VectorChangedEventArgs(CollectionChange change, uint index)
 	{
-		public VectorChangedEventArgs(CollectionChange change, uint index)
-		{
-			CollectionChange = change;
-			Index = index;
-		}
-
-		/// <inheritdoc />
-		public CollectionChange CollectionChange { get; }
-
-		/// <inheritdoc />
-		public uint Index { get; }
+		CollectionChange = change;
+		Index = index;
 	}
+
+	/// <inheritdoc />
+	public CollectionChange CollectionChange { get; }
+
+	/// <inheritdoc />
+	public uint Index { get; }
 }
