+++ src/Uno.Extensions.Reactive.UI/Presentation/Bindings/Collections/Views/FlatView.cs
@@ -8,7 +8,7 @@ using Windows.Foundation;
 using Windows.Foundation.Collections;
 using Uno.Extensions.Collections;
 using Umbrella.Feeds.Collections.Facades;
-using Umbrella.Presentation.Feeds.Collections._BindableCollection.Facets;
+using Uno.Extensions.Reactive.Bindings.Collections._BindableCollection.Facets;
 
 #if WINUI
 using CurrentChangingEventHandler = Microsoft.UI.Xaml.Data.CurrentChangingEventHandler;
@@ -24,7 +24,7 @@ using CurrentChangingEventArgs = System.ComponentModel.CurrentChangingEventArgs;
 using CurrentChangedEventHandler = System.EventHandler;
 #endif
 
-namespace Umbrella.Presentation.Feeds.Collections._BindableCollection.Views
+namespace Uno.Extensions.Reactive.Bindings.Collections._BindableCollection.Views
 {
 	/// <summary>
 	/// A view which flatten the items of a <see cref="IBindableCollectionViewSource"/>
