+++ src/Uno.Extensions.Reactive.UI/Presentation/Bindings/IListInput.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Linq;
+
+namespace Uno.Extensions.Reactive;
+
+/// <summary>
+/// 
+/// </summary>
+/// <typeparam name="T"></typeparam>
+public interface IListInput<T> : IListState<T>
+{
+	/// <summary>
+	/// The name of bindable property.
+	/// </summary>
+	public string PropertyName { get; }
+}
